refactor: RestaurantList UI요소와 도메인 로직 분리
